Change the bogon source from https://www.spamhaus.org to https://team-cymru.org
The former is a list of susspect IP ranges.
The latter is a list of IP unallocated addresses - they should not be see in routing.

I feel this more closely matches the purpose of not routing bogons.

This change means we don't need the jq dependency - it has been removed.

The IPv6 version of this file is bigger (about 150195 vs 78 lines).
As such, we need to adjust a sysctl to allow pf to handle that data.

As with period,conf, give the user the option to install that sysctl
to /etc/sysctl.conf.local